feat: redirect notification taps to notifcation tab
